Testing the command itself.
Used the eval File.read trick (from Rick, from script/runner ?) for these specs. The SystemExit rescue is so the exit call in the command doesn't kill the specs. This makes a lot of warnings for redefining constants, and I'd take care of that but I don't know what happened to remove_const.