Refactoring tag conversion.
Split out the tag logic in #mp3data to a separate method, #convert_tags,
which takes the entire tag structure and returns a new hash with pointers
to target (tag, tag2) and value. This is more in keeping with SRP and will
make it simpler to deal with tags that are separate for FLAC and combined
for MP3 (e.g. tracknumber = 1/tracktotal = 10 becomes TRCK = '1/10').