Defining Flac2mp3#get_flacdata.
This is almost identical to the previous incarnation as Flac2mp3.flacdata, and it uses the same method of knowing which fields should stay strings. This is just the start of the instance accessing what was previously a private class method, so the access restrictions are off (at least for now).